+++ src/frontend/app/features/cloud-foundry/services/cloud-foundry-organisation.service.ts
@@ -4,7 +4,8 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 import { filter, map, switchMap } from 'rxjs/operators';
 
-import { IApp, IOrganization, IPrivateDomain, IQuotaDefinition, IServiceInstance, ISpace } from '../../../core/cf-api.types';
+import { IServiceInstance } from '../../../core/cf-api-svc.types';
+import { IApp, IOrganization, IPrivateDomain, IQuotaDefinition, ISpace } from '../../../core/cf-api.types';
 import { EntityService } from '../../../core/entity-service';
 import { EntityServiceFactory } from '../../../core/entity-service-factory.service';
 import { CfUserService } from '../../../shared/data-services/cf-user.service';
