+++ src/frontend/app/shared/components/list/list-cards/meta-card/meta-card-base/meta-card.component.scss
@@ -1,6 +1,6 @@
 .meta-card {
   display: flex;
-  width: 100%;
-  height: 100%;
   flex-direction: column;
+  height: 100%;
+  width: 100%;
 }
