+++ src/frontend/app/shared/components/list/list-cards/meta-card/meta-card-item/meta-card-item.component.scss
@@ -29,10 +29,10 @@
 .meta-card-item-long-text {
   padding-top: 10px;
   .meta-card-item__value {
+    padding-bottom: 10px;
     text-align: start;
     white-space: unset;
     word-wrap: break-word;
-    padding-bottom: 10px;
   }
   .meta-card-item__key {
     display: none;
