+++ src/frontend/app/shared/components/list/list-types/cf-services/cf-service-card/cf-service-card.component.scss
@@ -8,9 +8,9 @@
       align-self: flex-end;
       img,
       mat-icon {
+        font-size: 48px;
         height: 48px;
         width: 48px;
-        font-size: 48px;
       }
     }
   }
